Finalized Sprites and Animations
Finished adding all of the sprites and animations into the game.
There are now 8 possible skins that the player and enemies can use.
The player has a white outline and the enemies have pink ones.
More items have also been added and there is a visual effect that happens when the player picks up an item.