Created an AudioManager and Settings Panel
Created an AudioManager to handle all the sounds in the app.
Added the main theme to the audio manager and set it to loop while playing.
Created a settings panel that can be accessed through the settings button in the main menu.
Volume for all "UI", "Game", and "Music" sounds can be adjusted in that panel.